+++ lib/make/component/choose_people_page.dart
