+++ lib/make/make_page.dart
@@ -1,27 +1,32 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:ore_kota/make/component/choose_people_page.dart';
 
-class QuestionDetail extends StatelessWidget {
-  const QuestionDetail({Key? key}) : super(key: key);
+class MakePage extends StatelessWidget {
+  const MakePage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("QuestionDetail"),
+        title: Text("makePage"),
       ),
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
             Text(
-              'QuestionDetail',
+              'Make',
             ),
           ],
         ),
       ),
       floatingActionButton: FloatingActionButton(
-        onPressed: () {},
+        onPressed: () {
+          Navigator.of(context).push(
+              MaterialPageRoute(builder: (context) => ChoosePeoplePage())
+          );
+        },
         tooltip: 'Increment',
         child: Icon(Icons.add),
       ),
