+++ lib/question/question_page.dart
@@ -1,32 +1,27 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:ore_kota/answer_page.dart';
 
-class AlertDialog extends StatelessWidget {
-  const AlertDialog({Key? key}) : super(key: key);
+class QuestionPage extends StatelessWidget {
+  const QuestionPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("dialog"),
+        title: Text("QuestionDetail"),
       ),
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
             Text(
-              'dialog',
+              'QuestionDetail',
             ),
           ],
         ),
       ),
       floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          Navigator.of(context).push(
-              MaterialPageRoute(builder: (context) => AnswerPage())
-          );
-        },
+        onPressed: () {},
         tooltip: 'Increment',
         child: Icon(Icons.add),
       ),
